Add a question to emphasise readers’ power & responsibility for planning
This is part of https://github.com/openaustralia/planningalerts/issues/992

“Should this go ahead?” reframes the planning feedback process to put
the citizen in the position of responsibility for how their area
changes.

It also encourages positive contributions as well: “I think this should
go ahead because ... ”. I think this is good because it makes
participating in local planning less defensive—you can also see that
there are possibilities to improve things as well as stop negative
changes.