Remind users that this is a public comment for everyone to see
This is to contribute to the sense of responsibility in users as they
participate in local planning. "public" emphasises that this is for everyone
to see, and that the user is part of a community who also have a say.

It's also a slight reminder to place nice, that people are listening and
you should be respectful.